move utils/blueprint to generator/base/internal
