+++ cli/environment-builder.mjs
@@ -27,7 +27,7 @@ import { fileURLToPath } from 'url';
 
 import { CLI_NAME, logger } from './utils.mjs';
 import { packageNameToNamespace } from '../generators/base/support/index.mjs';
-import { parseBlueprintInfo, loadBlueprintsFromConfiguration, mergeBlueprints } from '../utils/blueprint.mjs';
+import { parseBlueprintInfo, loadBlueprintsFromConfiguration, mergeBlueprints } from '../generators/base/internal/index.mjs';
 
 const __filename = fileURLToPath(import.meta.url);
 const __dirname = dirname(__filename);
